Easy function to execute the current file
This comes in handy when you're toying around or doing a demo. Calling
this function will execute the currently opened file using the correct
binary.

Would be cool to have this in its own plug-in, but I don't want to be
maintaining all the file types people might want this with. I only need
a few.